port to MSYS2/clang

+++ src/3rdParty/salomesmesh/src/StdMeshers/StdMeshers_ProjectionUtils.cpp
@@ -102,11 +102,11 @@ namespace HERE = StdMeshers_ProjectionUtils;
 namespace {
 
   static SMESHDS_Mesh* theMeshDS[2] = { 0, 0 }; // used for debug only
-  inline long shapeIndex(const TopoDS_Shape& S)
+  inline intptr_t shapeIndex(const TopoDS_Shape& S)
   {
     if ( theMeshDS[0] && theMeshDS[1] )
       return max(theMeshDS[0]->ShapeToIndex(S), theMeshDS[1]->ShapeToIndex(S) );
-    return long(S.TShape().operator->());
+    return intptr_t(S.TShape().operator->());
   }
 
   //================================================================================
